Fix locale dependent delay names (#6)
The technical names and display names of the delays are generated via `sprintf` calls which are locale dependent. This means that for example the feedback delay might be saved either as "fbdelay_0.1s" (point) or "fbdelay_0,1s" (comma).

The fix adds a string array next to the array of lengths and uses this string array when creating the names. This ensures that they will not be locale dependent and that they will always look the same.